Post Belt-Update bug fixes
Bug Fix: This was a misconception on my part. The "number of worlds" statistic for a system indicates the total number of gas giants, belts, and other worlds in the system. I was getting it to generate other worlds ON TOP OF gas giants, belts and the main world. Systems will look smaller from now on.

Bug Fix: Saving and loading systems was changing the position of stars and messing with other objects as a consequence. This was due to the seed for the system not being read correctly.

Bug Fix: The system generation was giving a different result almost every time, whether data was coming in from Traveller Map, the user was entering their own data, or the user was regenerating with the "current seed". This error was due to me not being careful with when and where I initiated the RNG with the correct seed.